--- src/engine/CMakeLists.txt
@@ -1,13 +1,0 @@
-add_library(engine STATIC 
-    matching_engine.cpp
-    order_book.cpp
-)
-
-target_include_directories(engine
-        PRIVATE
-        ${CMAKE_SOURCE_DIR}/include
-)
-
-
-target_precompile_headers(engine PRIVATE ${CMAKE_SOURCE_DIR}/include/pch.hpp)
-
