+++ src/matching-engine/CMakeLists.txt
@@ -1,8 +1,25 @@
-add_library(matching_engine STATIC matching_engine.cpp)
+add_library(matching_engine)
+
+target_sources(matching_engine
+    PRIVATE
+        ${CMAKE_CURRENT_SOURCE_DIR}/matching_engine.cpp 
+        ${CMAKE_CURRENT_SOURCE_DIR}/orderbook.cpp 
+        ${CMAKE_CURRENT_SOURCE_DIR}/sequencer.cpp 
+)
 
 target_include_directories(matching_engine
     PUBLIC
         ${CMAKE_SOURCE_DIR}/include
-        ${CMAKE_BINARY_DIR}/generated/include/
-        @ONLY
 )
+
+target_precompile_headers(matching_engine PRIVATE pch.h)
+
+#sequencer stuff
+find_package(concurrentqueue CONFIG REQUIRED)
+find_package(Threads REQUIRED)
+
+target_link_libraries(sequencer
+    PRIVATE concurrentqueue::concurrentqueue Threads::Threads
+)
+
+
